Person Access Filtering and Administrator
The routine “accessible_for_user” routine was not working as required and has been modified.
In doing this a new user type of “Administrator” has been added who is permitted to see all “Persons”.
The existing mongo migration script has been removed and replaced with a new script that creates the default administrator user.